chore: refactor project into plugins
