+++ src/assets/emissive_material.rs
