+++ src/assets/material_util.rs
