+++ src/ui/mod.rs
@@ -1,0 +1,21 @@
+use bevy::prelude::*;
+
+use self::{menu::MenuPlugin, scoreboard::ScoreboardPlugin};
+
+pub mod menu;
+pub mod scoreboard;
+
+pub struct UiPlugin;
+
+impl Plugin for UiPlugin {
+    fn build(&self, app: &mut App) {
+        app
+            // .add_plugin(MenuPlugin)
+            .add_plugin(ScoreboardPlugin)
+            .add_startup_system(setup_ui_camera);
+    }
+}
+
+fn setup_ui_camera(mut commands: Commands) {
+    commands.spawn_bundle(UiCameraBundle::default());
+}
